# API-WEB-SAUDE/src/swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "API de informações de Clínicas e Hospitais",
		"description": "Sobre a API de Clínicas e Hospitais",
		"contact": {
			"email": "ericspsantos@gmail.com"
		},
		"version": "1.0.0"
	},
	"servers": [
		{
			"url": "http://localhost:5000",
			"description": "API WEB SAÚDE local"
		},
		{
			"url": "https://api-web-saude.vercel.app/",
			"description": "API WEB SAÚDE hospedada"
		}
	],
	"paths": {
		"/login": {
			"post": {
				"summary": "Autenticar usuário",
				"tags": ["Usuarios"],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Login"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Autenticação bem-sucedida",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"token": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Login não autorizado",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Login"
								}
							}
						}
					},
					"500": {
						"description": "Erro interno do servidor"
					}
				}
			}
		},
		"/usuario/novo-usuario": {
			"post": {
				"summary": "Cadastrar Usuário",
				"description": "Essa rota será responsável por cadastrar um usuário",
				"tags": ["Usuarios"],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"requestBody": {
					"description": "Dados do Usuário",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Usuario"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Usuário cadastrado com Sucesso!",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Usuario"
								}
							}
						}
					},
					"400": {
						"description": "Campos inválidos ou nome muito curto",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Message": {
											"type": "string"
										},
										"Errors": {
											"type": "array",
											"items": {
												"type": "string"
											}
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Erro interno do servidor",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Message": {
											"type": "string"
										}
									}
								}
							}
						}
					}
				}
			}
		},
...
		"/usuario/deletar-usuario/{id}": {
			"delete": {
				"summary": "Deletar Usuário",
				"description": "Essa rota será responsável por deletar um usuário existente",
				"tags": ["Usuarios"],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						},
						"description": "ID do usuário a ser deletado"
					}
				],
				"responses": {
					"204": {
						"description": "Usuário deletado com sucesso"
					},
					"404": {
						"description": "Usuário não encontrado",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Message": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Erro interno do servidor",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Message": {
											"type": "string"
										}
									}
								}
							}
						}
					}
				}
			}
		},
		"/usuario/usuarios": {
			"get": {
				"summary": "Listar Usuários",
				"description": "Esta rota lista todos os usuários",
				"tags": ["Usuarios"],
				"responses": {
					"200": {
						"description": "Usuários listados com sucesso.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Usuario"
								}
							}
						}
					},
					"404": {
						"description": "Nenhum usuário encontrado",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Message": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Erro interno do servidor",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Message": {
											"type": "string"
										}
									}
								}
							}
						}
					}
				}
			}
		},
		"/usuario/usuario/nova-senha/{nome}": {
			"put": {
				"summary": "Alterar Senha do Usuário",
				"description": "Esta rota permite que o usuário altere sua senha",
				"tags": ["Usuarios"],
				"parameters": [
					{
						"name": "nome",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						},
						"description": "Nome do usuário"
					}
				],
				"requestBody": {
					"description": "Nova Senha do Usuário",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Senha"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Senha alterada com sucesso",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"mensagem": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Campos inválidos ou senhas não coincidem",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Message": {
											"type": "string"
										},
										"Errors": {
											"type": "array",
											"items": {
												"type": "string"
											}
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Erro interno do servidor",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Message": {
											"type": "string"
										}
									}
								}
							}
						}
					}
				}
			}
		},
		"/usuario/usuario/unidades-de-saude/": {
			"get": {
				"summary": "Listar Unidades de Saúde do Usuário",
				"description": "Esta rota lista todas as unidades de saúde associadas ao usuário",
				"tags": ["Usuarios"],
				"responses": {
					"200": {
						"description": "Unidades de Saúde listadas com sucesso.",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/UnidadeDeSaude"
									}
								}
							}
						}
					},
					"404": {
						"description": "Nenhuma unidade de saúde encontrada para o usuário",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Message": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Erro interno do servidor",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Message": {
											"type": "string"
										}
									}
								}
							}
						}
					}
				}
			}
		},
		"/usuario/usuario/{nome}": {
			"get": {
				"summary": "Pegar Usuário pelo Nome",
				"description": "Esta rota retorna informações do usuário pelo nome",
				"tags": ["Usuarios"],
				"parameters": [
					{
						"name": "nome",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						},
						"description": "Nome do usuário"
					}
				],
				"responses": {
					"200": {
						"description": "Usuário encontrado com sucesso.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Usuario"
								}
							}
						}
					},
					"404": {
						"description": "Usuário não encontrado",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Message": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Erro interno do servidor",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"Message": {
											"type": "string"
										}
									}
								}
							}
						}
					}
				}
			}
		},
		"/usuario/total-usuarios": {
			"get": {
				"summary": "Obter Total de Usuários",
				"description": "Esta rota retorna o total de usuários cadastrados",
				"tags": ["Usuarios"],
				"responses": {
					"200": {
						"description": "Total de usuários obtido com sucesso.",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"total": {
											"type": "integer"
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Erro interno do servidor",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "string"
										}
									}
								}
							}
						}
					}
				}
			}
		},
...
	},
	"components": {
		"schemas": {
			"Login": {
				"type": "object",
				"properties": {
					"email": {
						"type": "string"
					},
					"senha": {
						"type": "string"
					}
				},
				"required": ["email", "senha"]
			},
			"Usuario": {
				"type": "object",
				"properties": {
					"nome": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"senha": {
						"type": "string"
					}
				},
				"required": ["nome", "email", "senha"]
			},
			"Senha": {
				"type": "object",
				"properties": {
					"senha": {
						"type": "string"
					},
					"confirmarSenha": {
						"type": "string"
					}
				},
				"required": ["senha", "confirmarSenha"]
			},
			"UnidadeDeSaude": {
				"type": "object",
				"properties": {
					"nome": {
						"type": "string"
					},
					"endereco": {
						"type": "string"
					}
				},
				"required": ["nome", "endereco"]
			},
...
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	}
}